Split up normalize function by class and simplify.
